Restart the WS connection if no pongs are received
With this change, the server will actually respond to ping frames with a
corresponding pong frame.
We keep track of the outstanding pings on the client side and we restart
the connection if we have more than 3 outstanding pings.

The pongs are sent back automatically by the fastwebsockets library, but
to do so, the `WebsocketTunnelRead` struct needs to have access to the
`WriteHalf` of the WS connection, in order to send back the pongs. So we
now keep the `WriteHalf` in an `Arc<Mutex<_>>`.
The `PingState` that tracks the sent pings and received pongs is also
shared between the `WriteHalf` and `ReadHalf`.