Improve error message on websocket client close
Currently the server side prints the error message every time the client sends the Close frame.
Since the frame is sent on most connections, this leads to unneccary spam.
I haven't tested the http2 variant, but from looking at the code, it follows the same pattern with a different ErrorKind.

With the lower priority log level, this can be filtered easily.
The distinct messages also help with checking the logs by hand.