reverse_tunnel: Implement backoff retry upon remote server connection errors
Currently in reverse-tunnel mode, upon errors when client tries to establish
the connection to the remote server, 'run_reverse_tunnel' simply retries
every 1s.

This might be too agressive in certain scenarios, for example when
the tunnel server (or a load-blancer in front of it) is not yet ready
and returns 503. Another example is when the tunnel server does
not agree on the tunnel parameters.

Implement an exponential backoff for the delay between reconnect
attempts. The maximum reconnect delay is controlled by setting
"--reverse-reconnect-max-delay". The default is "1s" in order to keep
existing behavior.