don't require executor parameter
The fix for #433 (socket file leak) added an executor parameter. However, to get
the fix, users have to call executor.abort_all(). So the primary motivation for
this fix is so that all users get the fix even if they don't call abort_all().

The secondary motivation is to make the API nicer.